backend/requirements.txt
# VN30-Quantum Backend Requirements

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25

# HTTP Client
httpx==0.26.0

# AI
google-generativeai==0.3.2

# Data processing
numpy==1.26.3
pandas==2.1.4

# Market data
vnstock==3.0.9

# InfluxDB
influxdb-client==1.40.0

# Redis (optional)
redis==5.0.1

# Utils
python-dotenv==1.0.0
